chatView: replace old status tracking logic with the new one
First, the chatView needs to use the new UserStatusMonitor in order
to access the UserTracker class, so the old user tracking logic
needs to be replaced with the new one. The chatView only tracks
global status changes, hence the null parameter passed to the
UserTracker.watchRoomStatus() method. The handler used in the
case of global status changes is the _onStatusChanged() method.