userDetails: add notification support.
We want to be able to emit a notification whenever a
watched user comes online. The notification will be
emitted by the new userTracker module, but we also need
support for it in both the UserDetails and the UserPopover.
The UserDetails class needs to show an icon and a label
whenever the corresponding user is being watched, and
hide those when the user is not being watched. The new
notifications-enabled property is used to set the
visibility of the two widgets. This property is controlled
by the ToggleButton located in the UserPopover. We also
add a new icon that will be used in the UserDetails and
UserPopover classes.