userPopover: add notify button and action.
The ToggleButton (notifiButton) that is connected to the
notify action is added to the popover. We also bind its
visible property to its sensitive property, so that the
button is hidden whenever the action is disabled. Moreover,
the notifications-enabled property of the UserDetails are
bound to the active property of the notifyButton, so that
we can display the notifyLabel whenever the user is watched.