chatView: split out _createNickTag() and _onNickTagClicked()
The code that creates the tag is not complicated or long, but it's
definitely a lot better to split that out into the _createNickTag()
method, as the code is cleaner this way. The _onNickTagClicked()
method is responsible for creating the popover at the right position
upon clicking a nickTag. The popover is created only once for each
nickTag, but it is created on demand, i.e. when the _onNickTagClicked()
is run for the first time.