userDetails: add offline support
We want to be able to use the UserDetails class both when the
user is offline and online. The offline functionality of the
UserDetails class will be visible just in the UserPopover,
as there are no offline users displayed in the UserList. We
know which case we are in by the presence of the _user variable.
We also need to update the messageButton visibility. The _user
param is not passed at construction time, but is initialized
through a setter.