userTracker: add UserTracker class with local tracking only.
Polari needs a new module, a stand-alone way of tracking user
status across rooms. As the first part of the new UserTracker
module, this patch adds the local tracking functionality.
The local tracking part of the userTracker tracks status
changes that happen only inside a given room. Status changes
are handled with the status-changed::basenick detailed signal.
As a result, the tracking part that resided in the chatView
is now moved to the new userTracker module.