telepathyClient: Fix inconsistency on resume
When the :network-available property changes repeatedly in quick
succession (which apparently happens when resuming from suspend),
we can end up in an inconsistent state where an account has a
requested presence of OFFLINE, but is in fact fully connected.
Work around this by applying the same work-around as in the
previous commit.

https://bugzilla.gnome.org/show_bug.cgi?id=770751