application: Consider port for alternate server list
When we fail to connect to a network with multiple servers, we build
a list of alternate servers for retrying, excluding the original
server that already failed. Currently that filtering is based solely
on the server address, in the assumption that servers would not just
differ by port - however that's exactly the case for FreeNode, so
to not effectively disable connection retries there, we need to take
the port number into account as well.

https://bugzilla.gnome.org/show_bug.cgi?id=771217