main: Use structured log for logging
GLib gained the ability to log structured data rather than plain
strings this cycle, and with the addition of g_log_variant(), we
are now able to use it from introspection to add additional useful
information to our log messages when connected to the journal. And
when logging to the console, we are at least using the standard
GLib logging, which means no custom env variables anymore, support
for fatal-warnings etc.

https://bugzilla.gnome.org/show_bug.cgi?id=771218