telepathyClient: Fix acknowledging pending messages on close
Oh my, this is embarrassing: The actual method to remove all
pending messages is ack_all_pending_messages_*async*. Commit
5d07f45190 "worked" by triggering an exception before actually
leaving the channel, so the channel didn't pop up again because
it was never gone in the first place (though hidden from the UI).
It turns out that using the correct method doesn't fully work
either here, as the ::pending-message-removed handler relies on
the corresponding room which is gone when the signal is emitted.
We can work around this by acknowledging messages one-by-one and
running the handler manually, so do that to fix the original issue.

https://bugzilla.gnome.org/show_bug.cgi?id=768611