main: Turn on all messages when logging to journal
Unless turned on with the G_MESSAGES_DEBUG env variable, the default
log functions drop debug and info messages. While this behavior is
reasonable when writing to the console, the journal's log viewer
already filters messages by priority by default, so it makes more
sense to log everything in that case, as we won't have to ask people
to reproduce issues with debugging turned on.

https://bugzilla.gnome.org/show_bug.cgi?id=771218