room-list: Fix room name alignment
The visibility of the icon preceding a channel name is supposed to
depend on whether the channel represents a room or a private chat,
however as the icon is not excluded from show_all(), this only works
as intended for rooms joined later. While the icon is still empty
for the original set of rooms, the spacing between icon and label
results in names being aligned differently.

https://bugzilla.gnome.org/show_bug.cgi?id=771728