appNotifications: Add line wrap for simple notifications
When a command output is too long, they are not readable on smaller
screens. For example when you type /TOPIC command on a channel with
a very long topic, the notification is printed as one line and not
readable. Adding lineWrap to the notification labels properties makes
the label printed out as multiple lines and all of the output is
readable this way.

https://bugzilla.gnome.org/show_bug.cgi?id=721763