util: Split out match_identify_message() method
Matching an identify command is currently done by the room to emit
the ::identify-sent signal when the user entered the command, so
we can offer to save the password. We now want to mask the sent
password in the chat log, so split out the matching code into a
utility method.

https://bugzilla.gnome.org/show_bug.cgi?id=771888