userList: Postpone propagating height until content limit is set
For the user list popover to work as intended, we have to set both
:max-content-height - to limit the number of displayed entries - and
:propagate-natural-height - to actually request more than the minimum
height. As computing the former requires the rows' allocations, we
can only set it after the initial size request, however we currently
set the latter immediately. As a result, there's a short span where
the popover is set to request the full content height, which at best
leads to a visual glitch when the popover briefly flashes at full
height, and in the worst case will crash when the windowing system's
limits are exceeded. Fix this by only propagating the height once
the maximum content height is set as well.

https://bugzilla.gnome.org/show_bug.cgi?id=771557