build: Rename variable
We install both a (private) library and an executable called 'polari'.
For now, the executable is a symlink to a script rather than a compiled
binary, so we use the POLARI variable for the library's dependencies.
As we are about to change that, rename it to LIBPOLARI to free the name
for the executable.

https://bugzilla.gnome.org/show_bug.cgi?id=773293