build: Use a small C wrapper
Gjs' package module provides some convenience like automatically loading
resources and setting the correct WM_CLASS, however this requires apps
to follow the somehow problematic pattern of either using the App ID as
name of the main binary (which nobody does for the time being), or making
it a symlink to a script of that name. Unfortunately this pattern requires
special treatment in the build system (hello, meson), which then makes it
harder to support from IDEs (hello, builder). Instead, replace the current
launcher script with a small binary that runs the same code, but also
includes the required resources.

https://bugzilla.gnome.org/show_bug.cgi?id=773293