application: Add application style only once
We currently add the application style when initializing the main
window, which is fine as long as the window is truly unique.
However once we allow running the app in the background, this won't
be the case anymore - the window can be closed and a new one opened
at a later point. So to make sure the custom style is only added once,
add it during application startup instead.

https://bugzilla.gnome.org/show_bug.cgi?id=770750