entryArea: Exclude confirmation label from show_all()
The visibility is set programmatically (and controls the state or
visibility of other elements via property bindings), so make sure
to not show it accidentally via show_all().
(For now, the window is usually shown before rooms are added, but
this won't be the case anymore when we allow running the application
in the background).

https://bugzilla.gnome.org/show_bug.cgi?id=770750