flatpak: Bundle mission-control/idle
We currently rely on various telepathy services to be available on the
host - namely mission-control, idle and telepathy-logger. GNOME's builtin
chat integration means that mission-control and logger will likely be
available on GNOME systems, but idle is rarely if ever included in
default installs - and of course for non-GNOME systems, all bets are
off anyway.
To address that situation, bundle the required services in the sandbox.
They won't be visible to DBus' StartServiceByName method, but we'll
add code to launch the demons as necessary in a follow-up patch.

https://bugzilla.gnome.org/show_bug.cgi?id=777705