chatView: Don't access unset property
We attach a timestamp of the last activity to nick tags to decide
whether to show corresponding status messages or not. However
this monkey-patched property is unset if there hasn't been any
activity from the nick - make sure to handle that case properly
instead of using an undefined value in a computation.

https://bugzilla.gnome.org/show_bug.cgi?id=778793