joinDialog: Handle Escape in name entry
Now that the entry has been changed to a SearchEntry, it consumes
Escape key presses itself for the ::stop-search keybinding signal
rather than letting it bubble up to the dialog. Not being able to
dismiss the dialog via keyboard is inconvenient, so handle Escape
manually to clear the entry if it contains any text, or dismiss
the dialog otherwise.

https://bugzilla.gnome.org/show_bug.cgi?id=778797