serverRoomManager: Fix runtime warning
The callback to the GtkCellRendererToggle::toggled signal used different
names for the 'path' parameter in parameter list and function block,
whoops. It turns out that this error made the checkbox work correctly,
as we already handle the GtkTreeView::row-activated signal, which is
also emitted, to toggle the checked setting. Toggling a boolean twice
is clearly not what we want, so just remove the faulty signal handler.