util: Change URL matching (again)
Commit 094d7d1cc11 changed the URL regex to match any scheme under
the assumption that false positives would be rare. While this may
be true for schemes://, it turns out that trying to match colon-only
schemes generically commonly produces matches for random stuff that's
not a URL - someone not adding a whitespace after a mention, explaining
that server:port can be used to configure non-default ports, ...

So probably the best we can do is to keep allowing all schemes://, and
then whitelist a selected list of colon-only schemes. (Of course we
could go back to generating the list from the installed applications
when not running under flatpak, but let's try to keep the differences
between sandboxed and unsandboxed as small as possible).

https://bugzilla.gnome.org/show_bug.cgi?id=779449