room: Match on "login" to identify a nick as well
Some nick authentication bots are not only not called "NickServ", but
expect a command that is not "identify", sigh. Support some more of them
by matching on "login" as well as "identify" and remember the expected
command - as long as the parameters are still the same, the existing
NickServ support should then start to work for userserv and friends.

https://bugzilla.gnome.org/show_bug.cgi?id=778841