serverRoomList: Don't use filter text as room name to join
Now that the entry is used for filtering the room list, it is extremely
weird to also use its text as room name that is joined along the selected
rooms from the list - it means the user either has to enter the complete
room name (in which case selecting rooms from the list becomes pointless),
or remember to clear the entry after selecting a room.

Just remove the conflicting functionality for now, we will bring back
the ability to enter custom room names in a way that is compatible with
filtering soon.

https://bugzilla.gnome.org/show_bug.cgi?id=779960