roomManager: Add account parameter to lookupRoomByName()
Room names are only unique within a network, so we need an additional
account parameter to unequivocally identify a room. Being able to
/part rooms from any network may arguably be handy (yet unexpected),
but we really need to identify the correct room in the other case
where we match by room name:
The join dialog's room list presents already joined rooms as checked
but insensitive, so it is currently impossible to pick a room that
is joined on a different network.

https://bugzilla.gnome.org/show_bug.cgi?id=779906