serverRoomList: Provide an efficient way to join a single room
The ability to join multiple rooms at once is neat, but most useful
right after setting up a new network. After that, by far the most
common case is to join a single room, so it's worth making that
case as efficient as possible. Right now we require users to select
a room by selecting a row via mouse or keynav, and then activate
the join button. This is often slower than the pre-room-list dialog,
where a room was joined by hitting Enter after typing the (full)
room name - to improve on both current and old behavior, keep the
top result selected when filtering and toggle the selected row
when the entry is activated.

https://bugzilla.gnome.org/show_bug.cgi?id=779960