utils: Add updateTerms() helper
As we will soon add filtering to the server room list, it makes
sense to generalize how we normalize and split filter text into
terms. Add a utility method based on the code in ConnectionsList
for that purpose. As a bonus, the method's return value indicates
whether the terms actually changed, so callers can avoid unneeded
work (for example when the amount of stripped whitespace changed).

https://bugzilla.gnome.org/show_bug.cgi?id=779960