flatpak: Drop config-opts for polari
The runtime ships a single library in /usr/lib64, which is enough for
meson to use lib64 as default libdir. While nothing breaks with polari
installing to /app/lib64, it's quite ugly to split 64-bit libs like
that. That's why for consistency, we pass the --libdir option to meson,
and everything works out fine ... except that builder picks up the
options, but not the buildsystem, which fails as autotools' libdir
option expects an absolute path.

Meh. However we can still make flatpak-builder's meson builds use the
libdir we want without breaking builder's autotools build by patching
meson itself to never default to lib64 ...