app: Trim trailing _ instead of restoring original account
If a connection fails because the chosen nick is already in use, we
automatically retry with an appended underscore. But while we restore
the original name when the connection succeeds or we are giving up,
the underscores can stick around when polari is closed while still
connecting. And as the underscores are then considered part of the
original name when polari is next started, they tend to pile up after
a while. To counter that, assume that all trailing underscores aren't
part of the genuine nick and trim them instead of trying to restore the
original account name.

https://bugzilla.gnome.org/show_bug.cgi?id=781161