entryArea: Consider nickname changes when disconnected
Depending on whether the room is currently connected, we either use
the self-contact's :alias property or the account's :nickname one
for the nick button text. Either of those properties can change, but
we currently only update the text in the former case. Start tracking
changes to the :nickname property is well to make sure we properly
update the button while disconnected.