app: Reset nick on startup
Telepathy will remember the last used nickname and restore it when an
account gets connected. While not necessarily wrong, this conflicts
with us treating the nickname set from the popover as a temporary
change in contrast to the nickname configured in the properties,
so for us it makes more sense to reset the nickname at startup to
make sure it matches the configured one we are using to connect.

https://bugzilla.gnome.org/show_bug.cgi?id=781161