flatpak: Move flatpak manifests into flatpak/
The manifests are used both to build the stable and nightly flatpaks
on sdk.gnome.org, and to hack on Polari itself. For the former, adding
a separate manifest for the stable version to the master branch in
commit d723af25ca1 made a lot of sense, but unfortunately it has turned
out confusing for the latter case - users now need to pick the right
manifest for development, and of course it's the "wrong" one that
doesn't have any suffix and therefore looks more correct ...
We'll try to disentangle the two cases a bit to clear that up, starting
with moving the build configuration for the autobuilder out of the
toplevel.