telepathyClient: Only identify with nickname when necessary
Many NickServ bots only take a password parameter, while we currently
send the username along unconditionally to make sure that identification
is still possible when connecting with a nickname that doesn't match
the registered one.
In order to work both with bots that support an optional username
parameter and those that don't understand the parameter at all,
omit the username when it already matches the currently used nick.

https://bugzilla.gnome.org/show_bug.cgi?id=772915