serverRoomManager: Fix double-toggling of checkboxes
As activating a checkbox via keyboard navigation does not emit
the ::row-activated signal, we need to handle ::toggled as well.
However as clicking a row will emit both signals, we currently
end up toggling the checkbox twice in that case (that is, back
to its original state). Address this by only handling ::toggled
in response to keyboard events.

https://bugzilla.gnome.org/show_bug.cgi?id=782969