application: Give up on connection failure
We do the best we can to establish a connection, if the connection still
fails, then retrying endlessly is unlikely to turn up a different result.
Except that telepathy apparently disagrees, and keeps trying anyway;
stop it from doing that by explicitly setting the presence to offline,
until the network changes and it becomes worth checking again.

https://bugzilla.gnome.org/show_bug.cgi?id=784047