telepathyClient: Use GNetworkMonitor::network-changed signal
Unlike notify::network-available that is emitted when the availability
of a default route changes, ::network-changed is emitted on any network
changes, for example when previously unavailable ranges become available
after connecting to a VPN network. As this may well be the reason for a
previous connection failure, it makes sense for us to reconnect accounts
in response.

https://bugzilla.gnome.org/show_bug.cgi?id=784047