telepathyClient: Always include username when it's known to work
In addition to NickServ bots that don't support the username parameter
and those that optionally support it, there are also bots that require
it. To support those as well, store whether the user included the
username in the identify command - if it was, we know the parameter
is supported and we can always send it regardless of the currently
used nick.

https://bugzilla.gnome.org/show_bug.cgi?id=772915