telepathyClient: Replace non-standard replace parameter
String.replace only supports two parameters according to the standard.
Mozilla removed the non-standard flags parameter in version 47, so the
replacement is no longer global as intended, but only applies to the
first match. Luckily this is all we actually need in this case, but
let's still fix the code to do what it suggests.