entryArea: Use has_focus() method instead of property
GtkWidget has both a has_focus() method and a :has-focus property
(that we represent as 'has_focus'). So far gjs was able to deal with
the name overlap, but now trying to use the property will instead
test for the availability of the method. Just avoid the conflict
by switching to the method, which is unambiguous.