entryArea: ... and switch back to properties
In case where a method- and property name overlap, using the method
is less unambiguous than I thought - mozjs52-based gjs will only see
the method, while mozjs38-based gjs will only see the property. We
are in luck though, and the real property names contain dashes that
allow us to refer to the properties in a way that works for all gjs
versions.