lib: Add PolariTplImporter
Telepathy-logger uses GMarkup functions to escape markup when writing
logs, so for maximum compatibility we should use the same for reading.

That means the implementation has to happen in C, so add a new
importer class that imports the contents of an XML log file
into tracker.

Part-of: <https://gitlab.gnome.org/GNOME/polari/-/merge_requests/264>