lib: Add PolariMessage type
We have used a small message object to abstract differences between the
telepathy-glib and telepathy-logger message types ever since we started
to include backlogs. So far we just used a plain JS object, however we
will soon need to represent messages in C code as well, so add a new
boxed type that can replace the existing intermediate object.
This will promote the existing telepathy-logger dependency from runtime-
to build-time dependency, however this is temporary and we will soon get
rid of the dependency altogether.