connections: Add :favorites-only property
In order to provide a better onboarding experience, we will add a guided
initial setup mode based on the existing network/room configuration UI.
However we don't want to display the full list of predefined networks as
in the join dialog, or allow the creation of custom ones. To support this
use case, add a new contruct-only property to limit the list to favorite
networks rather than just prioritizing them over others.

https://bugzilla.gnome.org/show_bug.cgi?id=772458