joinDialog: Improve network selection via keyboard
In order to allow for keyboard-friendly network selection, we currently
activate the first one when the filter entry is activated. While this
is generally what we want, it's not very obvious that this will happen.
So instead, keep the first row selected while filtering, and activate
the selected entry when hitting enter.