initialSetup: Fix quit action
We will currently always open the main application window after the
initial setup window has been closed. This behavior is correct if
the window was completed or dismissed, but not when the app itself
was quit - cancel continuing to the main window in that case to
get the expected behavior.