appNotification: Fix account removal on quit
Application::prepare-shutdown is emitted when the window is destroyed,
after all its children has been destroyed as well - any signal connected
to an AppNotification has been disconnected by then, so even if we called
close as intended it wouldn't have any effect. So to get the intended
behavior, run close() when the notification is destroyed without being
closed before.