ci: Switch to rawhide
While F26 satisfies all our build dependencies (not least because
there's little to build), rawhide is much closer to the platform
we are actually developing for, in particular when it comes to gjs
and Spidermonkey.

Also rawhide's mozjs52-devel package now includes the js52 binary,
so we can run the basic syntax check added in commit 65a2099 during
continuous integration.