entryArea: Split out NickPopover
The popover for changing the nick is currently handled entirely as
part of the entryArea, which means we have to keep one popover for
every room around, even though they don't carry any room-specific
state and we never show more than one at any given time.
Splitting out the NickPopover decouples it from the entryArea,
which will allow us to share a single one in the future.