entryArea: Fix setting new nick immediately
The intended behavior when changing the nick via the popover is to
update the button label immediately, and only revert to the actually
used nick when the server doesn't react in a reasonable time (or
simply rejects the update).
However 911b95582f1 broke that, as the account's nickname property
changes right away , but isn't actually used to set the nick label
while connected. In that case the label is set from the connection's
self-contact, which will only pick up the new nick after it changed
on the server.
So to avoid resetting the newly requested nick immediately, only
consider changes to the nickname while disconnected.