flatpak: Pass --debug flag when launched from Builder
We strongly recommend newcomers to use Builder/Flatpak for development,
so it is important to make this work as best as possible out of the box.
As it is normal to have an IRC client running while hacking - and at least
for anyone hacking on Polari, it shouldn't be some other client - it would
be good to be able to test changes without continuously restarting the
regular instance.
Pass the newly added --debug flag when launched from Builder to enable this
for the recommended development workflow out of the box.

Fixes https://gitlab.gnome.org/GNOME/polari/issues/24