roomList: Add "Disconnect" to network context popover
It can be useful to temporarily disable an account - for instance
a work network while being on vacation - but we lost the ability
to do so from Polari itself when dropping the separate connections
dialog in commit 263ff952a3.

Re-expose the option by adding a "Disconnect" action to the network
context popover that mirrors the existing "Connect" one, and let
the actions control the account's :enabled property to preserve the
connection state over restarts.

https://bugzilla.gnome.org/show_bug.cgi?id=771889