accounts: Add PolariAccount:visible property
Ever since we removed the separate connections dialog in favor of
the room list header popovers in commit 263ff952a, it has not been
possible to enable or disable accounts from Polari itself. We would
like to re-expose that functionality again, but nowadays we also
use the disable mechanism to implement undo for account removals.
So in order to "free" the :enabled property, add a separate :visible
property that will be used to show or hide accounts from the UI.

https://bugzilla.gnome.org/show_bug.cgi?id=771889