build: Include prefix in libdir & co.
Previously, when absolute `libdir` path was needed, it was joined
with `prefix` on the spot. This meant the libraries used relative
`install_dir` when `libdir` option was not passed. This in turn
caused issues for NixOS’s patched Meson which needs absolute paths
to be able to use them in generated GIR files.

This commit initializes `bindir`, `libdir`, and `datadir` with
`prefix` in the top-level meson.build.