app: Clean up retry data of removed accounts
Telepathy's account IDs use a pattern of <CM>/<protocol>/<identifier>num,
which is mostly redundant for us: We only support idle/irc, and for
predefined networks the original nick (which ends up being the identifier
recorded in the ID) is the login username. So usually account IDs only
differ in the uniquifying number - starting from 0, telepathy will simply
pick the smallest number that hasn't been taken yet. That strategy means
that when removing an account, the next account that is created may well
use an identical ID. In case we have retry data (like alternative servers)
for the old account, we must make sure to not accidentally apply it to
the new one, so clean it on account removal.

Fixes https://gitlab.gnome.org/GNOME/polari/issues/27