telepathyClient: Fix missing notification on non GNOME
In non GNOME systems notifications are not shown when there is a private
message.

This happends because Polari creates custom notification in a way specific
to GNOME, this does not work on other desktop enviroments.

This fix is to send regular notifications that are supported by the Free Desktop standard when we
are not in GNOME.

Fixes: https://gitlab.gnome.org/GNOME/polari/issues/18