application: Defer loading of room lists on metered connections
We currently load all room lists immediately on startup, to hopefully
have them readily available when the user opens the join dialog. In
case the user doesn't use the dialog, this means we download the lists
unnecessarily, trading efficiency for convenience. This is a reasonable
trade-off where bandwidth usage isn't a concern (that is, network access
is sufficiently fast and cheap), but may very well be the wrong choice
otherwise.

At least in case where NetworkManager tells us that a connection is
metered, deferring loading the room lists until the join dialog is used
is the better option.

Fixes https://gitlab.gnome.org/GNOME/polari/issues/29