roomList: Disconnect popover signals
If the context popover is visible when the corresponding row is
destroyed - as happens in case the action is triggered from the
popover itself - then the popover's visibility will change one last
time before it is destroyed itself. Accessing the corresponding row
is already invalid in that case, so make sure to disconnect the
signals.

https://gitlab.gnome.org/GNOME/polari/issues/32