Added delay in spinner while reconnecting.
RoomListHeader spinner state changes very fast while reconnecting such that
when the state changes from 'error->connecting->error', it feels more like
'error->error' As @bastianilso wrote : 'The error can be reproduced when the
user is offline.' The same can also be seen when the IRC server is down or does
not exist.

The solution is a workaround to make the spinner in connecting state for at
least a second. When the connection status is of 'connecting', it tries to
swiftly change back to 'error' state. This has been stopped by placing a timeout
of at least a second (depending on how much time the iconStack spinner has been
in a 'connecting' state.

Original Issue : https://gitlab.gnome.org/GNOME/polari/issues/21