ui: Move app menu into window
GNOME is phasing out the app menu[0], so move its actions into a
hamburger menu inside the headerbar. Eventually we will remove
the app menu altogether, but for now we keep the 'quit' action
to mimic GNOME Shell's fallback menu, so that the action behaves
as expected (namely, quit the application instead of sending it
to the background if the corresponding option is set).

[0] https://wiki.gnome.org/Design/Whiteboards/AppMenuMigration

https://gitlab.gnome.org/GNOME/polari/merge_requests/53