flatpak: Disable API documentation
It is a bit pointless to install HTML documentation that we remove
from the final flatpak anyway, but so far we've built dependencies
from tarballs where the files are bundled and cannot be disabled.

But now gtk-doc broke the build with current org.gnome.Sdk//master,
so this has become more urgent - fix it by switching modules to
git and disable gtk-doc entirely.

https://gitlab.gnome.org/GNOME/polari/issues/64