room: Split out nick matching into a utility function
Nick highlighting is important to get right, which is why we try to
be smarter than simply matching for the current nick. As that makes
the code more complex and thus more prune to bugs, it is worth cove-
ring in unit testing. Splitting out the matching logic into a utility
function will allow us to do so without setting up a telepathy test
environment.

https://gitlab.gnome.org/GNOME/polari/issues/53