ci: Do not run fedora job automatically
The flatpak job is much better suited for regular CI, both because
it is what we recommend for development, and because it is based
on the nightly GNOME runtime (so it doesn't rely on released tarballs
for dependencies).

While making sure that we build on common distros is valuable as well,
it's something that we haven't flashed out yet. Until we do, don't
include the job in normal testing.