main: Split out ARGV conversion
By convention (and the example set by gjs-console), the javascript
code has access to a global ARGV variable that holds the actual
command line options (that is, ARGV[0] is not the executable from
argv[0]).

We currently handle that with some pointer arithmetic on argv itself,
but as we are about to do some semi-evil command line manipulation,
it makes sense to copy the relevant elements from the original argv
into a NULL-terminated string array.