build: Add support for 'snapshot' builds
Our CI now generates flatpak bundles on successful builds, which
are intended for conveniently testing proposed changes with just
a couple of clicks. However for now we still miss that goal, due
to two important restrictions:

 - the bundle uses the same ID/branch as the nightly
   build, so it can conflict or interfere with the
   installation

 - if Polari is already running in the session, the bundle
   has to be launched manually with --test-instance to work

We will be able to address the former by adjusting the CI script to
use a dedicated branch, but to address the latter, we need explicit
support in the app. Add that in form of a compile time option for
producing development snapshots, which injects --test-instance into
the command line.

https://gitlab.gnome.org/GNOME/polari/issues/61