build: Add support for 'snapshot' builds
Our CI now generates flatpak bundles on successful builds, which
are intended for conveniently testing proposed changes with just
a couple of clicks. However for now we still miss that goal, due
to two important restrictions:

 - the bundle uses the same ID/branch as the nightly
   build, so it can conflict or interfere with the
   installation

 - if Polari is already running in the session, the bundle
   has to be launched manually with --test-instance to work

In order to address that latter, add a 'snapshot' compile time
option that injects --test-instance into the command line.