ci: Set a cache key
Somehow the flatpak job keeps building all the dependencies more
often than not (despite all the modules being tagged to a specific
commit). I suspect gitlab throws the cache away when it encounters
a job with a different (in this case: no) path; if that is the
case, then specifying a key to use a non-default cache should
fix it.