Stop using TpChannelIface properties
As mozilla removed certain hooks from its API, gjs had to change the way
properties on introspected objects work - as a side effect, it now relies
more heavily on information provided by the typelib.

However TelepathyGLib doesn't expose most interfaces to introspection,
so their properties are now invisible to us. In theory this is easy to
fix - just pass the files in question to g-ir-scanner - but in practice
telepathy-glib hasn't seen any activity in years, so getting a patch
committed (and released in a tarball) is nigh impossible.

So instead, address this by using the corresponding accessor functions
instead of directly using the properties provided by affected interfaces.