main: Replace tedious regex with simpler string methods
We include fields about the code location where a message was logged
by throwing an error and parsing the resulting stack trace. The
regex we are currently using for that has at least two issues:

 - it grabs the column- instead of
   the line number

 - it doesn't handle some weird characters
   SpiderMonkey appends to some method names

Instead of increasing the complexity of the regex, use two simpler
split() operations to first retrieve the line information, and then
separate function and file.